refactor: split static data into modules
