+++ src/data/legend.ts
@@ -1,0 +1,7 @@
+export const LEGEND = [
+  { label: ">85", color: "bg-emerald-700" },
+  { label: "84–75", color: "bg-emerald-600" },
+  { label: "74–50", color: "bg-yellow-600" },
+  { label: "49–35", color: "bg-orange-600" },
+  { label: "34–25", color: "bg-red-600" }
+];
